added rsp_ functions
