+++ neurokit2/rsp/__init__.py
@@ -1,5 +1,8 @@
 """Submodule for NeuroKit."""
 
+from .rsp_findpeaks import rsp_findpeaks
 from .rsp_prepare import rsp_prepare
-from .rsp_process import Rsp
+from .rsp_rate import rsp_rate
+from .rsp_process import rsp_process
+from .Rsp import Rsp
